añadida funcion de ordenar
añadida funcionalidad de seleccionar tarea por titulo
el menu ahora solo se muestra si pulsamos enter

+++ TaskManager/src/taskmanager/view/TaskView.java
@@ -1,34 +1,37 @@
 package taskmanager.view;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 import taskmanager.data.TaskDataController;
 import taskmanager.data.TaskFileController;
-import taskmanager.tools.ToolsTaskManager;
 
 /**
  * class to implement the menu and the initial view
+ * 
  * @author Jose Mouriño
  *
  */
 public class TaskView {
-	
-	//get the path of the text file with data
+
+	// get the path of the text file with data
 	String path = TaskFileController.getPath();
+
 	
-	Scanner sc = new Scanner(System.in);
 	// loop to show the menu
 
 	public void showMenuView() {
-		//variable where we store the position of the task
-		int position;
-		//variable where we store a task
-		String task;	
+		
+		Scanner sc = new Scanner(System.in);
+		// variable where we store the position of the task
+		int position, newPosition;
+		// variable where we store a task
+		String task;
 
 		// variable where we store the value selected from the menu
 		int userOption = 0;
 		// data entry by keyboard
-		
+
 		do {
 
 			// print the menu by console
@@ -38,6 +41,7 @@ public class TaskView {
 			System.out.println("2. Listar Tareas");
 			System.out.println("3. Modificar Tarea");
 			System.out.println("4. Borrar Tarea");
+			System.out.println("5. Cambiar Tarea De Posicion");
 			System.out.println("0. Salir del programa");
 			System.out.print("Introduzca una opción válida: ");
 
@@ -46,7 +50,7 @@ public class TaskView {
 
 			// if the option is less than 0 or greater than 3, it is not a
 			// valid option
-			if (userOption < 0 || userOption > 4) {
+			if (userOption < 0 || userOption > 5) {
 				System.out.println("Opción no válida. Vuelva a escoger");
 			} else if (userOption != 0) {
 
@@ -61,35 +65,48 @@ public class TaskView {
 					task = sc.nextLine();
 					// save the task in the text file
 					TaskDataController.saveTask(task, path);
+					showInitialView();
 					break;
 				case LIST: // List Tasks
-					// get the list of tasks
-					List<String> tasks = TaskDataController.getTask(path);
-					// print the list of tasks if has elements
-					if (tasks.size() > 1) {
-						TaskDataController.printTask(tasks);
-					} else {
-						System.out.println("La lista de tareas está vacía");
-					}
-
+					showInitialView();
 					break;
 
 				case DELETE: // Delete Task
-					System.out.print("Introduzca la posicion de la tarea que desea borrar: ");
+					
 					// we obtain the position that we want to delete
-					position = ToolsTaskManager.getNumber();
+					position = selectTask(path);
 					// delete the task in position
-					TaskDataController.deleteTask(position, path);
+					boolean confirmationDelete = ToolsTaskManager.getConfirmation("borrar");
+					// get the confirmation of the delete
+					if(confirmationDelete == true) {
+						TaskDataController.deleteTask(position, path);
+						showInitialView();
+					}
 					break;
 
 				case MODIFY:
-					System.out.print("Introduzca la posicion de la tarea que desea modificar: ");
 					// we obtain the position that we want to modify
-					position = ToolsTaskManager.getNumber();
+					position = selectTask(path);
 					System.out.print("Introduzca la nueva tarea: ");
 					task = sc.nextLine();
-					// modify the task in position
-					TaskDataController.modifyTask(position, task, path);
+					boolean confirmationModify = ToolsTaskManager.getConfirmation("modificar");
+					// get the confirmation of the delete
+					if(confirmationModify == true) {
+						// modify the task in position
+						TaskDataController.modifyTask(position, task, path);
+						showInitialView();
+					}
+					
+					break;
+
+				case ORDER:
+					// we obtain the position that we want to modify
+					position = selectTask(path);
+					System.out.print("Ahora introduzca la nueva posicion para la tarea: ");
+					newPosition = ToolsTaskManager.getNumber();
+
+					TaskDataController.orderTask(position, newPosition, path);
+					showInitialView();
 					break;
 
 				}// switch
@@ -102,11 +119,13 @@ public class TaskView {
 		sc.close();
 
 	}
+
 	/**
 	 * method to show the initial view of the program with the list of task
 	 */
 	public void showInitialView() {
-		
+
+		Scanner sc = new Scanner(System.in);
 		// get the list of tasks
 		List<String> tasks = TaskDataController.getTask(path);
 		// print the list of tasks if has elements
@@ -116,12 +135,38 @@ public class TaskView {
 			System.out.println("La lista de tareas está vacía");
 		}
 		System.out.println("\n");
-		//show a message to go to the menu
-		System.out.printf("Pulse cualquier tecla para ir al menu principal: ");
-		if(sc.hasNextLine()) {
+		// show a message to go to the menu
+		System.out.println("Pulse la tecla enter para ir al menu.");
+		String enterkey = sc.nextLine();
+		if (enterkey.isEmpty()) {
 			showMenuView();
-			
+
 		}
+	}
+
+	public static int selectTask(String path) {
+
+		Scanner sc = new Scanner(System.in);
+		List<String> taskList = new ArrayList<>();
+		String title = "";
+		int position = 0;
+
+		// get the list of tasks
+		taskList = TaskDataController.getTask(path);
+
+		System.out.print("Introduzca la posicion o el titulo de la tarea que desea seleccionar: ");
+		if (sc.hasNextInt()) {
+			position = sc.nextInt();
+		} else {
+			title = sc.nextLine();
+		}
+
+		if (title != "") {
+
+			position = taskList.indexOf(title);
+		}
+		//sc.close();
+		return position;
 	}
 
 }
